+++ src/components/base/UpdateCheck.vue
@@ -1,5 +1,5 @@
 <template>
-  <q-dialog v-model="dialog" transition-duration="100">
+  <q-dialog v-model="dialog" transition-duration="100" persistent>
     <q-card style="width: 500px">
       <q-card-section class="flex justify-between">
         <h2 class="text-h6 q-my-none">