+++ src/components/snapshots/SnapshotsTable.vue
@@ -1,6 +1,13 @@
 <template>
   <div class="flex justify-between q-pa-md">
-    <new-snapshot :repository="repository" @reload="emit('reload')" />
+    <div class="flex items-center">
+      <new-snapshot :repository="repository" @reload="emit('reload')" />
+      <filter-state v-model="filter"
+                    :results-count="filterStateProps.resultsCount"
+                    :filtered-results-count="filterStateProps.filteredResultsCount"
+                    class="q-ml-md" />
+    </div>
+
     <filter-input v-model="filter" />
   </div>
 
@@ -8,8 +15,8 @@
            class="table-mono table-hide-overflow"
            dense
            row-key="name"
-           :columns="tableColumns"
-           :rows="snapshots"
+           :columns="columns"
+           :rows="filteredResults"
            :rows-per-page-options="DEFAULT_ROWS_PER_PAGE"
            :pagination="{sortBy: 'name'}">
     <template #body="{row}">
@@ -20,33 +27,14 @@
 
 <script setup lang="ts">
   import FilterInput from '../shared/FilterInput.vue'
-  import { ref } from 'vue'
   import { DEFAULT_ROWS_PER_PAGE } from '../../consts'
-  import { genColumns } from '../../helpers/tableColumns'
-  import { useTranslation } from '../../composables/i18n'
   import SnapshotRow from './SnapshotRow.vue'
   import NewSnapshot from './NewSnapshot.vue'
-  import { EsSnapshot } from '../../composables/components/snapshots/SnapshotRow.ts'
+  import { SnapshotsTableProps, useSnapshotsTable } from '../../composables/components/snapshots/SnapshotsTable.ts'
+  import FilterState from '../shared/FilterState.vue'
 
-  type SnapshotsTableProps = {
-    repository: string,
-    snapshots: EsSnapshot[]
-  }
-  defineProps<SnapshotsTableProps>()
+  const props = defineProps<SnapshotsTableProps>()
   const emit = defineEmits(['reload'])
 
-  const t = useTranslation()
-  const filter = ref('')
-  const tableColumns = genColumns([
-    { label: 'id' },
-    { label: t('snapshots.snapshots_table.table.headers.status'), field: 'state' },
-    { label: t('snapshots.snapshots_table.table.headers.start_time') },
-    { label: t('snapshots.snapshots_table.table.headers.end_time') },
-    { label: t('snapshots.snapshots_table.table.headers.duration') },
-    { label: t('snapshots.snapshots_table.table.headers.indices') },
-    { label: t('snapshots.snapshots_table.table.headers.successful_shards') },
-    { label: t('snapshots.snapshots_table.table.headers.failed_shards') },
-    { label: t('snapshots.snapshots_table.table.headers.total_shards') },
-    { label: '' }
-  ])
+  const { filter, columns, filteredResults, filterStateProps } = useSnapshotsTable(props)
 </script>