+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.4.0
+
+* for elasticsearch 7 and 8, query both `_template` and `_index_template` endpoints, fixes [#281](https://github.com/cars10/elasticvue/issues/281)
+* improve shards view
+* fix issue when multiple indices are selected and one is manually deleted
+* add shard recovery view, fixes [#272](https://github.com/cars10/elasticvue/issues/272)
+* adds chip indicating filtered tables. This should fix issues with users not noticing that the filter is still present.
+
 ## 1.3.1
 
 Change the build process for the linux desktop app. It now requires glibc `3.35`, so it should run distros using